requirements.txt
torch>=1.12
torchvision>=0.13
matplotlib>=3.5
gym[classic_control]>=0.26
pyyaml>=5.4
moviepy
imageio[ffmpeg]

